Refactor naming for ICLR20-Great related code
