+++ src/main/java/com/github/sergdelft/j2graph/Main.java
@@ -1,9 +1,11 @@
 package com.github.sergdelft.j2graph;
 
+import com.github.sergdelft.j2graph.iclr20great.ICLR20GreatDataGenerator;
+
 public class Main {
 
     public static void main(String[] args) {
-        DataGenerator dataGenerator = new DataGenerator();
-        dataGenerator.run();
+        ICLR20GreatDataGenerator ICLR20GreatDataGenerator = new ICLR20GreatDataGenerator();
+        ICLR20GreatDataGenerator.run();
     }
 }
